% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/closest.r
\name{closest}
\alias{closest}
\title{Closest relocation}
\usage{
closest(from, to = NULL, dt = c(-3600, 0), units = c("sec", "min", "hour",
  "day"), prefix = "to.", by = c("id", "burst"), range = c("[]", "[)",
  "(]", "()"), protect = NULL)
}
\arguments{
\item{from}{A ltraj object.}

\item{to}{A ltraj object, supposed to be different from
\code{from}.}

\item{dt}{A numeric of length 2, giving the boundaries of the
temporal window. See Details.}

\item{units}{A character string indicating the time units for
\code{dt}.}

\item{prefix}{A character string attached to the names of the
variables returned.}

\item{by}{Character. Either \code{"id"} to exclude the relocations
from the same individual in the computation of distances, or
\code{"burst"} to exclude relocations from the same burst.}

\item{range}{A character string indicating the range type, using a
representation with square and round brackets: a square bracket
means "inclusive" and a round bracket means "exclusive". For
example, \code{"[)"} includes the beginning, but not the end of
the interval.}

\item{protect}{A character string indicating other variables to
keep from \code{to} in \code{infolocs}.}
}
\value{
The same ltraj as \code{from}, with additional variables
giving x and y coordinates, the date, id and burst of the closest
relocation, together with the distance to this relocation.
}
\description{
Find closest relocation from ltraj objects.
}
\details{
Distances are computed between the relocations of two ltraj
objects within a temporal window. If only one ltraj is provided,
distances are computed for relocations of other individuals or
bursts (see parameter \code{by}).

\code{dt} allows to define the time window using \code{Inf}
(e.g. \code{dt = c(-Inf, 0)} corresponds to all previous
locations), as well as positive values (e.g. \code{dt = c(-3600,
3600)} corresponds to all locations from one hour before to one
hour after).
}
\note{
The function assumes that both ltraj are projected, and in
the same projection.
}
\examples{
data(puechcirc)

toto <- closest(puechcirc)
tail(infolocs(toto)[[3]])

tata <- closest(puechcirc, dt = c(-3, 2), units = "hour", by = "burst")
tail(infolocs(tata)[[3]])

titi <- closest(puechcirc, dt = c(-5, -2), units = "hour", protect = "abs.angle")
tail(infolocs(titi)[[3]])
}
\author{
Mathieu Basille \email{basille@ase-research.org}
}

