% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{hab}
\alias{hab}
\alias{hab-package}
\title{Habitat selection and movement analyses}
\description{
A set of functions related to habitat selection and movement
analyses. Also includes a few patches for functions from the
\code{adehabitatXY} package family. For a list of documented
functions, use \code{library(help = "hab")}
}
\author{
Mathieu Basille \email{basille@ase-research.org}
}

