% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dl}
\alias{dl}
\alias{ld}
\title{Quick Conversion of Objects of Class ltraj from and to
Dataframes}
\usage{
dl(x, strict = TRUE)

ld(ltraj, strict = TRUE)
}
\arguments{
\item{strict}{Logical. Whether to use the regular
\code{\link[adehabitatLT]{ld}} or \code{\link[adehabitatLT]{dl}}
functions, which enforce more verifications.}
}
\description{
Faster versions of \code{\link[adehabitatLT]{ld}} and
\code{\link[adehabitatLT]{dl}}.
}
\details{
In \code{\link[adehabitatLT]{ld}}, \code{strict = FALSE} can be up
to 10 times faster, but assumes that the \code{ltraj} is well
structured (i.e. not modified by the user). In
\code{\link[adehabitatLT]{dl}}, \code{strict = FALSE} can be up to
20 times faster, but assumes that the trajectory parameters in the
data frame (x/y increments, angles, etc.) are still valid (e.g. no
locations have been removed).
}
\examples{
data(puechcirc)
puechcirc ## class ltraj

## ld
df1 <- adehabitatLT:::ld(puechcirc)
df2 <- ld(puechcirc, strict = FALSE)
all.equal(df1, df2)
## Note a difference in row names:
attr(df1, "row.names")
attr(df2, "row.names")

## dl
all.equal(dl(df2), adehabitatLT:::dl(df2))
dl(df2, strict = FALSE)
## Comparison regarding 'strict'
all.equal(dl(df2), dl(df2, strict = FALSE))
## Differences in row.names (numeric in regular 'dl', characters using
## 'strict = FALSE') + NAs in R2n (for a reason, 'puechcirc[[2]]'
## starts by a sequence of missing values, but has several 'R2n'
## values. As a result, 'strict = FALSE' keeps the 'R2n' values)
}
\author{
Modified by Mathieu Basille
\email{basille@ase-research.org}
}
\seealso{
See \code{\link[adehabitatLT]{ld}} for further details on
the function and all available arguments.
}

