% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeSeq.r
\name{makeSeq}
\alias{makeSeq}
\alias{print.summarySeq}
\alias{summarySeq}
\title{Define sequences of a trajectory}
\usage{
makeSeq(x, gap = 1, units = c("hours", "mins", "secs", "days", "weeks"),
  na.omit = TRUE, name = "seq")

summarySeq(x, name = "seq")

\method{print}{summarySeq}(x, ...)
}
\arguments{
\item{x}{A ltraj object.}

\item{gap}{The maximum time interval between two successives steps
before starting a new sequence (note that the time interval needs
to be strictly greater than \code{gap} for a new sequence to
start).}

\item{units}{The unit of \code{gap} (default is \code{hour}).}

\item{na.omit}{Logical, whether to remove missing locations to
form the sequence (default).}

\item{name}{A character string indicating the column name of the
sequence to be stored in infolocs.}

\item{x}{A ltraj object.}

\item{name}{A character string indicating the column name of the
sequence computed by \code{makeSeq}.}

\item{x}{An object of class \code{summarySeq}.}
}
\value{
For \code{makeSeq}, a ltraj with a new variable
\code{name} in infolocs, with the sequence number for each
location, in the form of an integer series for each burst.

For \code{summarySeq}, an object of class \code{summarySeq}.
}
\description{
Define sequences of a trajectory based on the chronology of steps,
allowing for gaps.
}
\examples{
data(puechcirc)

puechcirc <- makeSeq(puechcirc, gap = 15, units = "mins", name = "seq15")
infolocs(puechcirc, which = "seq15", simplify = TRUE)
summarySeq(puechcirc, "seq15")

puechcirc <- makeSeq(puechcirc, gap = 25, units = "mins", name = "seq25")
summarySeq(puechcirc, "seq25")
}
\author{
Mathieu Basille \email{basille@ase-research.org}
}

