% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rdSteps}
\alias{rdSteps}
\title{Draw random steps}
\usage{
rdSteps(x, nrs = 10, distMax = Inf, others = FALSE, simult = FALSE,
  reproducible = FALSE)
}
\arguments{
\item{x}{A ltraj object.}

\item{nrs}{The number of random steps to draw for each observed step
(default = 10).}

\item{distMax}{Only draw step lengths and turning angles using steps
shorter than this threshold. Default is \code{Inf}, i.e. all steps are
kept.}

\item{only.others}{Logical, draws step lengths and turning angles from
all other individuals, excluding the current one.}

\item{simult}{Logical, whether to draw step lengths and turning angles
simultaneously, i.e. both measurements come from a single observed
step, instead of being drawn independently.}

\item{reproducible}{Logical,}
}
\value{
A data frame, with new columns \code{case} (1 for observed
steps and 0 for random steps) and \code{strata} (an common integer for
paired observed and random steps).
}
\description{
Draw random steps from a ltraj object using empirical distributions of
step lengths and turning angles.
}
\details{
Note that 1) only complete steps are kept (i.e. steps characterized by
start and end points, and turning angle (i.e. relative angle to the
previous step); and 2) the information stored in infolocs is transfered
to all random steps within a strata (i.e. it assumes it is the same for
all steps within a strata).
}
\examples{
## Load the data
data(puechcirc)

## Simple example to check the distributions of step lengths and turning
## angles
bla <- rdSteps(puechcirc)
boxplot(bla$rel.angle ~ bla$case)
boxplot(bla$dist ~ bla$case)

## Reproducibility
head(rdSteps(puechcirc, reproducible = TRUE))
head(rdSteps(puechcirc, reproducible = TRUE))
}
\author{
Mathieu Basille \email{basille@ase-research.org}
}

