% Generated by roxygen2 (4.0.1): do not edit by hand
\name{subset.ltraj}
\alias{subset.ltraj}
\title{Subsetting a ltraj}
\usage{
\method{subset}{ltraj}(x, subset, rec = FALSE, ...)
}
\arguments{
\item{x}{A ltraj object.}

\item{subset}{Logical expression indicating elements or rows to keep:
missing values are taken as false.}

\item{rec}{Logical, whether to recompute the ltraj parameters or not
(default = FALSE).}
}
\value{
A ltraj object.
}
\description{
Return subsets of a ltraj which meet conditions (over the descriptive
parameters of the ltraj or its infolocs).
}
\examples{
## Load puechcirc data and add some infolocs:
data(puechcirc)
info <- list(data.frame(A = rnorm(80), B = runif(80), C = rpois(80,
    1)), data.frame(A = rnorm(69, 10), B = runif(69, 10, 11),
    C = rpois(69, 10)), data.frame(A = rnorm(66, -10), B = runif(66,
    -10, -9), C = rpois(66, 100)))
## Watch the row names:
info <- mapply(function(x, y) {
    row.names(x) <- row.names(y)
    return(x)
}, info, puechcirc, SIMPLIFY = FALSE)
infolocs(puechcirc) <- info

## Different subsets:
(xsub1 <- subset(puechcirc, dist > 200, rec = FALSE))
(xsub2 <- subset(puechcirc, C == 3, rec = TRUE))
(xsub3 <- subset(puechcirc, C == 3, rec = FALSE))
}
\author{
Mathieu Basille \email{basille@ase-research.org}
}
\seealso{
See \code{\link[adehabitatLT]{which.ltraj}} to identify the
relocations fullfilling a condition.
}

