% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/infolocs.r
\name{infolocs}
\alias{infolocs}
\title{Infolocs of an Object of Class ltraj}
\usage{
infolocs(ltraj, which, by = c("burst", "id", "none"), simplify = FALSE)
}
\arguments{
\item{by}{Character, replaces \code{perani}. Either \code{"burst"}
(identical to \code{perani = FALSE}), \code{"id"} (identical to
\code{perani = TRUE}), or \code{"none"} to return only one element
for all bursts together.}

\item{simplify}{Logical. If a single variable is requested, should
the function return 1-column data frames (\code{FALSE}, default)
or vector (\code{TRUE}).}
}
\description{
Modified version of \code{\link[adehabitatLT]{infolocs}} that
returns \code{NULL} if \code{infolocs} exists, but \code{which} is
not a colum of it (note that \code{adehabitatLT::infolocs} already
returns \code{NULL} if there is no infolocs).
}
\examples{
## Load puechcirc data, and add some random infolocs:
data(puechcirc)
infolocs(puechcirc)
info <- list(data.frame(A = rnorm(80), B = runif(80), C = rpois(80,
    1)), data.frame(A = rnorm(69, 10), B = runif(69, 10, 11),
    C = rpois(69, 10)), data.frame(A = rnorm(66, -10), B = runif(66,
    -10, -9), C = rpois(66, 100)))
## Watch the row names:
info <- mapply(function(x, y) {
    row.names(x) <- row.names(y)
    return(x)
}, info, puechcirc, SIMPLIFY = FALSE)
infolocs(puechcirc) <- info
puechcirc

## Parameters "by" and "simplify"
infolocs(puechcirc, "A", simplify = TRUE)
infolocs(puechcirc, "A", by = "id", simplify = TRUE)
infolocs(puechcirc, "A", by = "none", simplify = TRUE)

## Try to retrieve the column `toto`:
infolocs(puechcirc, "toto")
}
\author{
Modified by Mathieu Basille
\email{basille@ase-research.org}
}
\seealso{
See \code{\link[adehabitatLT]{infolocs}} for further
details on the function and all available arguments.
}

