% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makeSeq}
\alias{makeSeq}
\alias{print.summarySeq}
\alias{summarySeq}
\title{Define sequences of a trajectory}
\usage{
makeSeq(date, id = NULL, gap = 1, units = c("hours", "mins", "secs",
  "days", "weeks"))

summarySeq(seq, id = NULL)

\method{print}{summarySeq}(x, ...)
}
\arguments{
\item{date}{A \code{POSIXt} object, chronologically ordered (by
individal if \code{id} is provided).}

\item{id}{A vector giving the id of each step (must be the same length
as \code{date}). If not provided, the function assumes all steps come
from a single individual.}

\item{gap}{The maximum time interval between two successives steps
before starting a new sequence.}

\item{units}{The unit of \code{gap} (default is \code{hour}).}

\item{seq}{An vector giving a sequence in a trajectory, as given by
\code{makeSeq}, or as a series of integers.}

\item{x}{An object of class \code{summarySeq}.}
}
\value{
For \code{makeSeq}, a numeric vector, with the sequence number
for each \code{date}, in the form of an integer series for each
individual.

For \code{summarySeq}, an object of class \code{summarySeq}.
}
\description{
Define sequences of a trajectory based on the chronology of steps,
allowing for gaps.
}
\examples{
## Dummy data:
(steps <- data.frame(date = Sys.time() + c(1:10, 12, 14, 20:25) *
    3600, id = c(rep("toto", 4), rep("tata", 9), rep("toto",
    5))))

## 1-hour gap, all steps considered from a single individual:
makeSeq(steps$date)
## 2-hour gap, all steps considered from a single individual:
makeSeq(steps$date, gap = 2)
## 1-hour gap, individual id considered:
steps$seq <- makeSeq(steps$date, steps$id)
steps

## Note that `seq` uses a series starting from 1 for each individual. Uses
## something like this to have unique seq IDs among individuals:
steps$seqid <- paste(steps$id, steps$seq, sep = "-")
steps

## Summary of the sequence, individual id considered:
summarySeq(steps$seq, steps$id)
}
\author{
Mathieu Basille \email{basille@ase-research.org}
}

