% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/na.omit.ltraj.r
\name{na.omit.ltraj}
\alias{na.omit.ltraj}
\title{Handle Missing Values in Objects of Class 'ltraj'}
\usage{
\method{na.omit}{ltraj}(object, rec = TRUE, complete.steps = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{ltraj}.}

\item{rec}{Logical, whether to recompute descriptive parameters of the
trajectory (in particular dx, dy, and angles). Use \code{FALSE} with
care.}

\item{complete.steps}{Logical, whether to keep only complete steps,
i.e. steps characterized by start and end points, and turning angle
(i.e. relative angle to the previous step).}

\item{...}{Further arguments not used.}
}
\value{
A ltraj object, without missing locations.
}
\description{
\code{na.omit} removes missing locations from a \code{ltraj} object.
}
\examples{
data(puechcirc)
puechcirc
na.omit(puechcirc)
## Note the direct call to hab:::
hab:::na.omit.ltraj(puechcirc, rec = FALSE, complete.steps = TRUE)
}
\author{
Mathieu Basille \email{basille@ase-research.org}
}

