% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rdSteps.r
\name{rdSteps}
\alias{rdSteps}
\title{Draw random steps}
\usage{
rdSteps(x, nrs = 10, rand.dis = NULL, only.others = FALSE,
  simult = FALSE, distMax = Inf, strata = NULL, mask = NULL,
  reproducible = FALSE)
}
\arguments{
\item{x}{A ltraj object.}

\item{nrs}{The number of random steps to draw for each observed step
(default = 10).}

\item{rand.dis}{The random distributions for step lengths and turning
angles to use. If \code{NULL} (default), it uses \code{x} as a basis;
otherwise, another \code{ltraj} object must be provided, or a
data.frame with columns "dist", "rel.angle" and "id".}

\item{only.others}{Logical, draws step lengths and turning angles from
all other individuals, excluding the current one.}

\item{simult}{Logical, whether to draw step lengths and turning angles
simultaneously, i.e. both measurements come from a single observed
step, instead of being drawn independently.}

\item{distMax}{Only draw step lengths and turning angles using steps
shorter than this threshold. Default is \code{Inf}, i.e. all steps are
kept.}

\item{strata}{Character. How to name the stratas. In all cases, a new
column \code{strata} is created. If \code{NULL}, a series of integer
from 1 to the total number of steps (on every individuals) is used; if
\code{"row.names"}, the row names are used; otherwise the column
provided is used.}

\item{mask}{A \code{SpatialPolygons} used to redraw steps that end
outside of the polygon. Use with caution: there is absolutely no
guarantee that the function is not caught in an endless loop (warnings
are issued if there are long or very long loops). Note that it is
assumed that steps are in the same projection as the mask (if it is not
the case, project the mask first using
\code{\link[rgdal]{spTransform}}).}

\item{reproducible}{Logical. If \code{TRUE}, results are made
reproducible with the use of a seed, otherwise new random step lengths
and turning angles are sampled at each call.}
}
\value{
A data frame, with new columns \code{case} (1 for observed
steps and 0 for random steps) and \code{strata} (a unique value for
each set of paired observed and random steps).
}
\description{
Draw random steps from a ltraj object using empirical distributions of
step lengths and turning angles.
}
\details{
Note that 1) only complete steps are kept (i.e. steps characterized by
start and end points, and turning angle (i.e. relative angle to the
previous step); and 2) the information stored in infolocs is transfered
to all random steps within a strata (i.e. it assumes it is the same for
all steps within a strata).
}
\examples{
## Load the data
data(puechcirc)
##'
## Simple example to check the distributions of step lengths and turning
## angles
bla <- rdSteps(puechcirc)
boxplot(bla$rel.angle ~ bla$case)
boxplot(bla$dist ~ bla$case)

## Reproducibility and alternative random distributions

## 1) Default: using the same ltraj for the random distributions:
bla <- rdSteps(puechcirc, reproducible = TRUE)

## 2) Explicitly use the same ltraj for the random distributions:
bli <- rdSteps(puechcirc, rand.dis = puechcirc, reproducible = TRUE)

## Check that 2) is the same as 1)
all.equal(bla, bli)

## 3) Explicitly uses random distributions in a data.frame:
rand <- subset(ld(puechcirc), !(is.na(x) | is.na(dx) | is.na(rel.angle)) &
    dist <= Inf, select = c("dist", "rel.angle", "id"))
blo <- rdSteps(puechcirc, rand.dis = rand, reproducible = TRUE)

## Check that 3) is the same as 1)
all.equal(bla, blo)
}
\author{
Mathieu Basille \email{basille@ase-research.org}
}

